refactor: :recycle: ユーザー検索関連のコードを切り出し
