+++ test/feature/definition/application/definition_id_list_state_test.mocks.dart
@@ -633,19 +633,6 @@ class MockUserProfileRepository extends _i1.Mock
       ) as _i10.Future<_i5.UserProfileDocument>);
 
   @override
-  _i10.Future<_i5.UserProfileDocument?> searchUserProfileByPublicId(
-          String? publicId) =>
-      (super.noSuchMethod(
-        Invocation.method(
-          #searchUserProfileByPublicId,
-          [publicId],
-        ),
-        returnValue: _i10.Future<_i5.UserProfileDocument?>.value(),
-        returnValueForMissingStub:
-            _i10.Future<_i5.UserProfileDocument?>.value(),
-      ) as _i10.Future<_i5.UserProfileDocument?>);
-
-  @override
   _i10.Future<void> updateUserProfile(_i16.UserProfile? userProfileForWrite) =>
       (super.noSuchMethod(
         Invocation.method(
